+++ plugins/org.ietr.preesm.codegen.xtend/ecore-gen/org/ietr/preesm/codegen/xtend/model/codegen/impl/CodegenFactoryImpl.java
@@ -464,7 +464,7 @@ public class CodegenFactoryImpl extends EFactoryImpl implements CodegenFactory {
 
   /**
    * <!-- begin-user-doc --> <!-- end-user-doc -->
-   *
+   * 
    * @generated
    */
   @Override