+++ release_notes.md
@@ -31,6 +31,7 @@ PREESM Changelog
 * Update external dependencies to [3.3.0,4.0.0);
 * Update external Graphiti to [1.10.0,1.0.0);
 * Update external DFTools to [1.9.0,2.0.0);
+* All objects in the codegen model are now able to hold comments;
 
 ### Bug fix
 * Fix issue #62